Fix signedness warnings in gdk-pixbuf/gdk-pixdata.c
gdk-pixbuf/gdk-pixdata.c: In function ‘gdk_pixdata_deserialize’:
gdk-pixbuf/gdk-pixdata.c:244:21: error: comparison of integer expressions of different signedness: ‘guint’ {aka ‘unsigned int’} and ‘int’ [-Werror=sign-compare]
   if (stream_length < pixdata->length - GDK_PIXDATA_HEADER_LENGTH)
                     ^
gdk-pixbuf/gdk-pixdata.c: In function ‘gdk_pixbuf_from_pixdata’:
gdk-pixbuf/gdk-pixdata.c:476:23: error: comparison of integer expressions of different signedness: ‘gint32’ {aka ‘const int’} and ‘unsigned int’ [-Werror=sign-compare]
       pixdata->length < pixdata->height * pixdata->rowstride - GDK_PIXDATA_HEADER_LENGTH)
                       ^

+++ gdk-pixbuf/gdk-pixdata.c
@@ -241,7 +241,7 @@ gdk_pixdata_deserialize (GdkPixdata   *pixdata,
     return_invalid_format (error);
 
   /* deserialize pixel data */
-  if (stream_length < pixdata->length - GDK_PIXDATA_HEADER_LENGTH)
+  if (stream_length < (guint) (pixdata->length - GDK_PIXDATA_HEADER_LENGTH))
     return_pixel_corrupt (error);
   pixdata->pixel_data = (guint8 *)stream;
 
@@ -473,7 +473,7 @@ gdk_pixbuf_from_pixdata (const GdkPixdata *pixdata,
 
   if (encoding == GDK_PIXDATA_ENCODING_RAW &&
       pixdata->length >= 1 &&
-      pixdata->length < pixdata->height * pixdata->rowstride - GDK_PIXDATA_HEADER_LENGTH)
+      pixdata->length < (gint) (pixdata->height * pixdata->rowstride - GDK_PIXDATA_HEADER_LENGTH))
     {
       g_set_error_literal (error, GDK_PIXBUF_ERROR,
                            GDK_PIXBUF_ERROR_CORRUPT_IMAGE,
